Parse arguments with strong types and add strings
Currently, the parsing of arguments is left up to each instruction,
making the behavior inconsistent. Required arguments can be omitted and
are substituted with 0. Opcodes that don't take arguments ignore any
arguments that are given. `storestr` mangles a char argument, because it
converts it to the decimal string representation of its code unit.

=> Scan arguments into tokens and check their types before invoking the
function for the opcode. Represent this within the TypeScript type
system using discriminated unions, so it is strongly typed. Throw on
malformed input.

It relies heavily on splitting lines by exactly one space, making it
brittle to other whitespace characters like tab or repeated spaces. At
the start and end of the line, Unicode whitespace may be used due to
`String.prototype.trim`.

=> Consistently use /\s/ while parsing and create a "-quoted space
token, so that strings with spaces can still be represented. Single-word
strings are still allowed.